// frontend/src/lib/api/models/match.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Paris Sportif API
 * API for football predictions and sports betting analysis
 * OpenAPI spec version: 1.0.0
 */
import type { MatchStatus } from './matchStatus';
import type { Team } from './team';

export interface Match {
  id: number;
  external_id?: string;
  home_team: Team | string;
  away_team: Team | string;
  competition: string;
  competition_code: string;
  match_date: string;
  status: MatchStatus;
  /** @nullable */
  home_score?: number | null;
  /** @nullable */
  away_score?: number | null;
  matchday?: number;
}